Project restructure, Tile component
