+++ src/components/Tile.jsx
@@ -1,0 +1,9 @@
+const Tile = (props) => {
+  return (
+    <div id={props.id} className={props.class}>
+      {props.children}
+    </div>
+  )
+}
+
+export default Tile;
