+++ src/layouts/MainSection.jsx
@@ -1,4 +1,4 @@
-import HomePage from './HomePage'
+import HomePage from './HomePage.jsx'
 
 const MainSection = (props) => {
   if (props.show === "homepage") {
