+++ src/style/components/BurgerButton.css
@@ -1,5 +1,3 @@
-@import "../global/Variables.css";
-
 @media (min-width: 1000px) {
     .burger-button {
         display: none;
