+++ src/style/components/Tiles.css
@@ -1,0 +1,23 @@
+@import "../../global/Variables.css";
+
+#tile-recently-played {
+    width: 100%;
+    height: 30rem;
+}
+
+#tile-recently-played, .basic-tile {
+    background-color: var(--color-side-bar);
+}
+
+.basic-tile {
+    height: 15rem;
+    width: 20rem;
+}
+
+.basic-tile:hover {
+    background-color: var(--color-hover) /*!important*/;
+}
+
+.basic-tile:active {
+    background-color: var(--color-selection) /*!important*/;
+}
