+++ src/style/layout/HomePageLayout.css
@@ -24,24 +24,3 @@
     }
 }
 
-#tile-recently-played {
-    width: 100%;
-    height: 30rem;
-}
-
-#tile-recently-played, #tile-settings, #tile-about, #tile-help {
-    background-color: var(--color-side-bar);
-}
-
-.basic-tile {
-    height: 15rem;
-    width: 20rem;
-}
-
-.basic-tile:hover {
-    background-color: var(--color-hover) !important;
-}
-
-.basic-tile:active {
-    background-color: var(--color-selection) !important;
-}
